// pronti-frontend-main/front/src/app/modulos/trabajador/form/form-trabajador.enum.ts
import {SiNoEnum} from '../../../enums/si-no.enum';

export enum FormTrabajadorEnum {
  // nombre = 'nombre',
  usuario = 'idUsuario',
  modalidadContrato = 'modalidadContrato',
  estadoCivil = 'estadoCivil',
  genero = 'genero',
  grupoSanguineo = 'grupoSanguineo',
  nivelEstudios = 'nivelEstudios',
  profesion = 'profesion',
  direccionDomiciliaria = 'direccionDomiciliaria',
  estadoFamiliar = 'estadoFamiliar',
  // codigoSectorial = 'codigoSectorial',
  numeroAfiliacion = 'numeroAfiliacion',
  discapacidad = 'discapacidad',
  tipoDiscapacidad = 'tipoDiscapacidad',
  sueldo = 'sueldo',
  bonificacion = 'bonificacion',
  movilizacionEspecial = 'movilizacionEspecial',
  componenteSalarialUnifi = 'componenteSalarialUnifi',
  retencionesJudiciales = 'retencionesJudiciales',
  polizaPersonal = 'polizaPersonal',
  aporteIess = 'aporteIess',
  aporteIessAsumido = 'aporteIessAsumido',
  decimosAnio = 'decimosAnio',
  impuestoRenta = 'impuestoRenta',
  fondoReservaIess = 'fondoReservaIess',
  utilidades = 'utilidades',
  pagoFondosReservaMes = 'pagoFondosReservaMes',
  impuestoRentaPatron = 'impuestoRentaPatron',
  vacaciones = 'vacaciones',
  beneficios = 'beneficios',
  pagoDecTercerCuartoMes = 'pagoDecTercerCuartoMes',
  acumuladoAportePersonalIess = 'acumuladoAportePersonalIess',
  acumuladoDecimoTercero = 'acumuladoDecimoTercero',
  acumuladoFondosReserva = 'acumuladoFondosReserva',
  acumuladoVacaciones = 'acumuladoVacaciones',
  acumuladoImpuestoRenta = 'acumuladoImpuestoRenta',
  utilidadCargas = 'utilidadCargas',
  tiempoParcial = 'tiempoParcial', //
  factorParcial = 'factorParcial',//
  pasante = 'pasante',//
  reingreso = 'reingreso',//
  fechaReingreso = 'fechaReingreso',//
  fechaIngreso = 'fechaIngreso',
  provDecimoTercero = 'provDecimoTercero',
  decimoCuarto = 'decimoCuartoA',
  provDecimoCuarto = 'provDecimoCuarto',
  provFondosReserva = 'provFondosReserva',
  provVacaciones = 'provVacaciones',
  provAportePatronal = 'provAportePatronal',
  cuentaContableNombre = 'cuentaContableNombre',
  idAgencia = 'idAgencia'

}